Update README.md
Deleted the last part of the first sentence of the testing section. Added the steps to go to the county of the users choice. Added the steps to zoom in or out of the Google Map and to view Google Maps in a new browser tab or in Google Maps if it's installed